+++ src/dk/aau/cs/verification/TAPNComposer.java
@@ -70,7 +70,7 @@ public class TAPNComposer implements ITAPNComposer {
 		hasShownMessage = false;
 
 		
-		double greatestWidth = 0,
+		int greatestWidth = 0,
 			   greatestHeight = 0;
 		if (this.guiModels != null) {
 			for (TimedArcPetriNet tapn1 : model.activeTemplates()) {
@@ -195,7 +195,7 @@ public class TAPNComposer implements ITAPNComposer {
 		}
 	}
 
-	private void createPlaces(TimedArcPetriNetNetwork model, TimedArcPetriNet constructedModel, NameMapping mapping, DataLayer guiModel, double greatestWidth, double greatestHeight) {
+	private void createPlaces(TimedArcPetriNetNetwork model, TimedArcPetriNet constructedModel, NameMapping mapping, DataLayer guiModel, int greatestWidth, int greatestHeight) {
 		int i = 0;
 		for (TimedArcPetriNet tapn : model.activeTemplates()) {
 			DataLayer currentGuiModel = null;
@@ -246,7 +246,7 @@ public class TAPNComposer implements ITAPNComposer {
 		}
 	}
 
-	private void createTransitions(TimedArcPetriNetNetwork model, TimedArcPetriNet constructedModel, NameMapping mapping, DataLayer guiModel, double greatestWidth, double greatestHeight) {
+	private void createTransitions(TimedArcPetriNetNetwork model, TimedArcPetriNet constructedModel, NameMapping mapping, DataLayer guiModel, int greatestWidth, int greatestHeight) {
 		int i = 0;
 		for (TimedArcPetriNet tapn : model.activeTemplates()) {
                         
@@ -318,7 +318,7 @@ public class TAPNComposer implements ITAPNComposer {
 		}
 	}
 	
-	private ArcPath createArcPath(DataLayer currentGuiModel, PlaceTransitionObject source, PlaceTransitionObject target, Arc arc, double offsetX, double offsetY) {
+	private ArcPath createArcPath(DataLayer currentGuiModel, PlaceTransitionObject source, PlaceTransitionObject target, Arc arc, int offsetX, int offsetY) {
 		Arc guiArc = currentGuiModel.getArcByEndpoints(source, target);
 		ArcPath arcPath = guiArc.getArcPath();
 		int arcPathPointsNum = arcPath.getNumPoints();
@@ -338,7 +338,7 @@ public class TAPNComposer implements ITAPNComposer {
 		return newArcPath;
 	}
 
-	private void createInputArcs(TimedArcPetriNetNetwork model, TimedArcPetriNet constructedModel, NameMapping mapping, DataLayer guiModel, double greatestWidth, double greatestHeight) {
+	private void createInputArcs(TimedArcPetriNetNetwork model, TimedArcPetriNet constructedModel, NameMapping mapping, DataLayer guiModel, int greatestWidth, int greatestHeight) {
 		int i = 0;
 		for (TimedArcPetriNet tapn : model.activeTemplates()) {
                         
@@ -402,7 +402,7 @@ public class TAPNComposer implements ITAPNComposer {
 		}
 	}
 
-	private void createOutputArcs(TimedArcPetriNetNetwork model, TimedArcPetriNet constructedModel, NameMapping mapping, DataLayer guiModel, double greatestWidth, double greatestHeight) {
+	private void createOutputArcs(TimedArcPetriNetNetwork model, TimedArcPetriNet constructedModel, NameMapping mapping, DataLayer guiModel, int greatestWidth, int greatestHeight) {
 		int i = 0;
 		for (TimedArcPetriNet tapn : model.activeTemplates()) {
 			DataLayer currentGuiModel = null;
@@ -457,7 +457,7 @@ public class TAPNComposer implements ITAPNComposer {
 		}
 	}
 
-	private void createTransportArcs(TimedArcPetriNetNetwork model, TimedArcPetriNet constructedModel, NameMapping mapping, DataLayer guiModel, double greatestWidth, double greatestHeight) {
+	private void createTransportArcs(TimedArcPetriNetNetwork model, TimedArcPetriNet constructedModel, NameMapping mapping, DataLayer guiModel, int greatestWidth, int greatestHeight) {
 		int i = 0;
 		int nextGroupNr = 0;
 		for (TimedArcPetriNet tapn : model.activeTemplates()) {
@@ -574,7 +574,7 @@ public class TAPNComposer implements ITAPNComposer {
 
 	
 	
-	private void createInhibitorArcs(TimedArcPetriNetNetwork model, TimedArcPetriNet constructedModel, NameMapping mapping, DataLayer guiModel, double greatestWidth, double greatestHeight) {
+	private void createInhibitorArcs(TimedArcPetriNetNetwork model, TimedArcPetriNet constructedModel, NameMapping mapping, DataLayer guiModel, int greatestWidth, int greatestHeight) {
 		int i = 0;
 		for (TimedArcPetriNet tapn : model.activeTemplates()) {
